config/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.8.0
Django==4.2.7
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.20.1
idna==3.4
install==1.3.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
maturin==1.3.2
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.2.0
networkx==3.2.1
numpy==1.26.3
openai==1.3.5
packaging==23.2
param==2.0.1
parso==0.8.3
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.21.12
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dotenv==1.0.0
pytorch-lightning==2.1.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stability-sdk==0.8.5
stack-data==0.6.3
sympy==1.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchmetrics==1.2.1
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
yarl==1.9.4
